Remove examples showing how to configure RSpace as global file source
There are two ways to enable file sources: globally or as file source templates. File source templates allow each user to define their own file sources, providing their credentials in the process; the newly defined file sources are only shown in the list of file sources after they have been set up. Global file sources are always shown on the list of file sources; credentials may be provided on the user preferences page.

Everything that is "personal", for example Dropbox or your own S3 buckets should be enabled using file source templates. Everything that is public AND requires no credentials (or equivalently, it is meant to be accessible to all users of the instance, as it may happen in a private lab) should also be enabled using file source templates. The remaining may be configured as global file sources.

RSpace falls in the first category, thus it makes sense to discourage configuring it as a global file source.