[25.0] Rerun workflows for the correct version/instance
In `GET /api/workflows/{encoded_workflow_id}/download`, for the case when `instance = True`, we were correctly obtaining the stored workflow.

However, then in `WorkflowContentsManager.workflow_to_dict` we were not considering the case that when `instance = True`, we need to not return the `latest_workflow.workflow_to_dict`, and instead return the specific workflow instance which has been specified.

We did handle the case when `version` is explicitly provided here, but not instance.

Fixes https://github.com/galaxyproject/galaxy/issues/20223