Attempt to replace duplicate datasets if we can
The conditions are:
- the replacement dataset exists in the same object store that the upload job would use
- the replacement dataset exists in a history that the user owns
- the replacement dataset has a HDA of the same datatype
- the dataset hash matches
- the replacement dataset is not purged or deleted and in OK state
- the same transform is applied

For now this is most interesting for the galaxy-live demo, where we're
going to upload datasets from dropbox and have the job cache kick in.

However, the deduplication is not tied to uploads, and we could consider
opening this up to public datasets and/or datasets that users have
access permissions on.