wait for add items button to enable; no need to switch to create tab
The test would fail because the add items to collection button's disabled/enabled state was not being detected, and hence, we would manually switch tabs and the items would not even have been added to the list yet. Now, we wait for the button to enable, and then adding the items automatically switches tab to Create.