use text only collection element; fix rename bug; add rename modal
This aims to fix all of the issues here: Fixes https://github.com/galaxyproject/galaxy/issues/20441

Using the default text-only value for the slot in `FormSelectMany` works best as it prevents layout shifting on hover/focused states.

This commit also fixes a bug with the `renameElement` function, which would create a duplicate of the renamed element (1 in `inListElements`, 1 in `workingElements` upon rename).

This adds a modal where the user can choose to rename elements if they'd like, instead of renaming the elements inside the multi select; which is not intuitive and can also interfere with clicking an element to select it.