fix duplicates being pushed to `uploadedDatasets` from upload values
In the case of remote file uploads, for each upload value, we have the entire set of uploads in the `model.outputs` object, which would cause the same dataset to appear multiple times in the `uploadedHistoryItems` object.

For other, non-remote uploads, we typically only have the matching output for the upload value in the `model.outputs` object.