Add ``/api/datasets/{dataset_id}/extra_files/raw/{filename:path}``
We need this for the vitessce visualization, that will build urls based
on the path of the current url. In practice, to get the file expected at `obs/.zattrs` the
visualization will build
https://usegalaxy.org/api/datasets/f9cad7b01a4721358531a44ecfb1aff6/display/obs/.zattrs?filename=A/1/bdbd683b-1870-4c11-ae29-b35ff0be59b2.adata.zarr,
because it reasonably assumes the query parameter is independent of the
path.
With this change we can access the file in question via
https://usegalaxy.org/api/datasets/f9cad7b01a4721358531a44ecfb1aff6/extra_files/raw/A/1/bdbd683b-1870-4c11-ae29-b35ff0be59b2.adata.zarr/obs/.zattrs.