Propagate cached job output replacement to copies (made by database operation tools)
Should fix the issue we had during Galaxy live, where we had a step in
which we picked the first dataset collection element from a mapped over
collection whose outputs where not yet replaced by the job cache.

The effect of this is that the extracted dataset's dataset was deleted,
and all downstream jobs where waiting for this dataset to become
terminal.