Display the Galaxy Citation within the Export Tool References List
This commit implements a new feature to display a user-defined Galaxy Citation
within the "Export Tool References List". By default, this includes
a reference to the NAR paper, but this can be changed at runtime
via a galaxy.yml edit or by rebuilding the client with an updated
config_schma.yml. This feature was proposed and discussed here:
https://github.com/galaxyproject/galaxy/issues/19362

Summary of changes:
- Define a new configuration property for citation_bibtex with a reference to the current main galaxy paper.
- Inject that as a citation within the Export Tool Refernences to diplay alongside other tools used in the current history.
- Uses the Cite() package to format in bibtex or APA as needed (as done for all other tools too)
- Updates the docs, and schema as needed to provide the default value and description