Fix using multiple data inputs with deferred datasets & Pulsar.
We really have to eliminate that input_datasets dictionary - this is very hacky. Maybe we can take a pass at rewriting Pulsar staging after we have stronger typing in the tool state branch for this stuff. Now we're putting multiple copies of an HDA into the deferred object structure - so I am now caching them during materialization so we don't materialize them more than once - this was probably needed anyway regardless since an input can be supplied to a tool more than once.