fix stored invocation check in invocation state view
The initial purpose of adding this was to prevent the page from reloading when you switch tabs. We set the boolean as soon as we know that the invocation exists in the store, and with this, when we change tabs there isn't a refresh every time.

But we mustn't return from the watcher at this point, we must continue polling.