Change Advanced Tool Search to a Tool Discovery View
We've discussed the idea of moving the advanced tool search out of the tool panel, and into the "search results" view for a while. It makes sense because the advanced search, which searches via the backend, is decoupled from the client search that applies to the panel itself and is entirely client side.

Therefore, we've moved the filters to the center panel in the rich tools list view, and more work will be done on making this view the best place to find ideal tools for our users' analyses.

This implements part of the plans in https://github.com/galaxyproject/galaxy/issues/17286