add jest for complex case where `propItems` update externally
An example of this is:

You load a few invocations in the scroll list by scrolling; say 20 loaded out of 120 for example.

You run a workflow separately. The total invocations are 121 now. While the `ScrollList` component has the correct current count 21 (since `propItems` will also be updated via the `invocationStore`), without the `props.adjustForTotalCountChanges` adjustment in `totalItemCount`, it would still show total count as 120 since all it knows is what the loader loaded so far. This is why we have that count adjustment in the computed variable.

Then, when you scroll again, the loader will be called, and this fetch will update the total count via the backend, meaning the prop state and local state will be synced fully.