reset `formConfig.errors` if there is no `refreshRequest`
Also consider `this.validationInternal` as errors.

The clearing of `formConfig.errors` on no refresh request seems to work because in `FormDisplay`, we seem to be returning that boolean as a param when there is a change in value that would possibly return errors from the backend.

And of course, `this.validationInternal` contains errors found by the client for individual inputs and is therefore, an essential check to disable run.