Speed up ImplicitCollectionJobs.job_list
Minor tweak, for 100 items this goes from 500ms to 80ms and I assume
this becomes more significant the more jobs there are.

No changes to the API, we should probably paginate this and just return
relevant fields, but this should address the 12s load time on the
backend when you click on a 8000+ job step.