Attempt to fix transient API test failures.
I don't really think this will fix the problem - this seems like a deep bug in Galaxy to me but it might be worth attempting. The two things I did here is wait on the testing initial condition (the job search can find the job before we delete the dataset) in case there is a race condition of some sort and then use the more percise dataset deletion route in case there is a clash between collections and datasets. I think we defer to the dataset so this second thing probably won't help but I think the more percise endpoint simplifies the test and ensures this isn't the problem.