only show overview tab for subworkflow steps in invocation view
This was initially to fix the bug where for a subworkflow (and expandable workflow runs in the invocation grid), we were showing all tabs/nav items and clicking them was changing the route to the subworkflow's (or expanded run's) detailed view on that tab (unexpected behavior when on the parent invocation view or the grid).

However, this actually makes sense to only show the overview tab because if a user wishes to view other items such as Report, Export etc, they would be better off viewing that workflow's detailed view in another tab (as guided by the overview tab via the link in `SubworkflowAlert.vue` added here).