Probably fix transiently failing test.
There is a sleep in the job now so the test runner is much less likely to not reach the delete before the job is completed naturally. Also on the off chance the issue is part of the dataset being committed before the message - there is a wait loop now.