[25.1] Prevent duplicate `current_history_json` calls
Uses a unique ID to prevent the resource watcher from creating multiple timeouts in a row. Also places the resource watcher in the history store as opposed to the `watchHistoryProvided` file which I have removed here.

The way to test this bug was to wait for the history to stop polling, and then click on the reload button in the panel many times, which would do as many duplicate calls (because it would create several different timeouts).