[25.1] Add a client side rate limiter
Adds a rate limiter middleware to the `openapi-fetch` client side API factory.

For now, the maximum number of requests are defined in this `rateLimiter` utility file (and so is the window duration, retry delay for 429s and max retries). Maybe this could be defined in a config file?

Also, we do not handle max retries for 429 responses (Retry-After) for now.