only allow one draft at a time, and update it if it is the target version
A limitation of the release-drafter (as I understand it) is that it
only updates the latest draft, whether it is the dev draft at that
point or the release draft, hence, resulting in duplicate release
drafts if a PR against a release branch is merged.

Here, we've added a step that checks and expects only 1 draft release
at a time. This is cleaner because: while the only draft is `dev`,
we only update that draft. Once, during the release process, `dev`
becomes a `release_` branch, we do not create yet another draft
for dev and instead just update the latest release branch draft,
until the release will be published and then dev will start getting
its own draft notes.